Move TaskEventViewSetMixin to mixins module
