+++ django-rgd/rgd/rest/base.py
@@ -1,33 +1,5 @@
-from drf_yasg.utils import swagger_auto_schema
 from rest_framework import mixins, viewsets
-from rest_framework.decorators import action
-from rest_framework.response import Response
 from rgd.rest.mixins import BaseRestViewMixin
-
-
-class TaskEventViewSetMixin:
-    """
-    Mixin for any TaskEventMixin model.
-
-    Provides the `status` action to retrieve the status of a
-    model's task event.
-
-    TODO: This should use a serializer instead.
-    """
-
-    @swagger_auto_schema(
-        method='GET',
-        operation_summary='Check the status.',
-    )
-    @action(detail=True)
-    def status(self, *args, **kwargs):
-        obj = self.get_object()
-        data = {
-            'pk': obj.pk,
-            'model': obj.__name__,
-            'status': obj.status,
-        }
-        return Response(data)
 
 
 class ReadOnlyModelViewSet(
